% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{reviews}
\alias{reviews}
\title{Movie Review Data}
\format{A list with element containing a review. The element name can be used to identify
the review 'number' and whether the review was positive/negative.}
\description{
Movie Review Data
}
\references{
http://www.cs.cornell.edu/people/pabo/movie-review-data/
}

