% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{reviews}
\alias{reviews}
\title{A set of 2000 movie reviews parsed by Pang and Lee (ACL, 2004), originally
gathered from the IMDB archive of the rec.arts.movies.reviews newsgroup}
\format{Character vector of length 2000. Each element of that vector contains a single movie review.}
\source{
\url{http://www.cs.cornell.edu/people/pabo/movie-review-data/}
}
\usage{
reviews
}
\description{
A set of 2000 movie reviews parsed by Pang and Lee (ACL, 2004), originally
gathered from the IMDB archive of the rec.arts.movies.reviews newsgroup
}
\keyword{datasets}

