% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{reviews}
\alias{reviews}
\title{Movie Review Data}
\format{A character vector where each element contains a review.
The element name can be used to identify the review 'number'
and whether the review was positive/negative.}
\source{
\url{http://www.cs.cornell.edu/people/pabo/movie-review-data/}
}
\description{
Movie reviews that are often used in sentiment analysis. Note this is v2.0
of the polarity data.
}

